+++ crates/vala/vala-bifrost-redux/src/tables/traces/links.rs
@@ -1,6 +1,6 @@
 use arrow::datatypes::Field;
 
-use crate::tables::fields::{fixed_binary, uint32, utf8};
+use crate::tables::fields::{fixed_binary, int64, utf8};
 use crate::tables::{
     CorrelationPolicy, DeclaredIndex, DomainTable, IndexKind, PayloadClass, SortKey,
 };
@@ -22,9 +22,9 @@ impl DomainTable for LinksTable {
             fixed_binary("linked_trace_id", 16, false),
             fixed_binary("linked_span_id", 8, false),
             utf8("trace_state", true),
-            uint32("flags", false),
+            int64("flags", false),
             utf8("attributes", true),
-            uint32("dropped_attributes_count", false),
+            int64("dropped_attributes_count", false),
         ]
     }
